Extract TemplateControl dropdown to component
